requirements.txt
aiogram>=3.0.0
aiosqlite>=0.17.0
APScheduler>=3.9.1
python-dotenv>=0.20.0